Make the bootstrap API easier to use
An enum encodes the meaning of its values. This removes the need to store the
value of the connection type flag in a variable to make it clear.

Configure extra profiling in `DatadogConfig` so that services to not have to
implement configuration support for this.

This is a breaking change.