feat: Exposed GetSpanFromContext so that dev can trace, hard to trace func
We started with #191.

The opaque abstraction although shields the devs from details of datadog,
seems to be creating more issue.

We start with supporting tracing for function with 1 argument, and then
more might follow.

Rather than that, an alternative solution is to expose GetSpanFromContext
which will allow the devs to get the span and then trace to their heart's
content. But, this also means, the abstraction is now leaky.

I have also introduced noopSpan and noopContext so that devs do not have
to nil check the span everywhere.