feat: Add Datadog hook for go-logger lib
If the hook is added to the go-logger lib, and then a log-message is logged with `logger.WithContext(ctx).Info("my-message")`, then any datadog trace-information in the context will be added to the log-message's fields.

Note: This commit will fail until go-logger has released a new version, with the HookEntry-struct.