chore: remove repo's release-drafter configuration (#264)
We are already using GitHub's magic repo `.github` to have a shared config
across all repos, and because this file was not diverging from
the old config and was just redundant. Removing it.

Individual repos should only host their own version of
`release-drafter.yaml` when they want to diverge from the common
way of labeling and changelog generator that is defined
in magic repo `.github`.