fix: gromtrace panic when nil config is provided
`gormtrace.Open` call `gorm.Open` which accepts vardic options arguments.
And if config passed is nil, `gorm.Open` accepts at is a valid `Option`
and tries to call `Apply` on it an panic.

There isn't much we can do about it except for passing in a new
`Config` when gormCfg is nil.