fix: Avoid race condition
Discards any errors from the cancel() call in our local goroutine. The errors will be reported to the depending system ifor when they call cancel() themselves, as we expect them to. This change avoids a race condition where another cancel() call will read the cancelErr value at the same time as we overwrite it (with the same value! The value is also being set on line 55).