refactor: Simplify metric-options
Use the same options-type (both the struct, and the functional argument) for the global metric options and the per-metric options. This means that we can use the same "WithSampleRate", "WithTag" etc for the global and the per-metric options. It also allows for configuring a custom errorhandler for a single metric (e.g. where it is imperative that we are able to report the metric)

Also, if an error happens during resolving of per-metric options, then the error-handler will be used (global, or successfully set with per-metric options)

This PR is stacked on #524, and will not prevent us from completing #527