feat: Add a logging adapter for go-logger
This change moves the Datadog logging adapter from `go-logger` to
`go-datadog-lib`. With the recent changes to `dd-trace-go` `go-logger` should not
have a dependency on `dd-trace-go` so this is better implemented in
`go-datadog-lib` that must a dependency on `go-logger`. Nothing should use this
adapter directly outside of `go-datadog-lib` therefor it is implemented as a
internal package.