fix: Verfiry that the database driver is a pointer
This changes verifies that the database driver is a pointer, passing the driver
as a value leads to dd-trace-go using its default service name e.g.
`mysql.mysqldriver` instead of the service name from `DD_SERVICE` or the one
provided through `database.WithServiceName`. This leads to poor observability.

The previous example code passed the driver as a value while the unit tests
passed the driver as a pointer, leading to confusion.